% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_data}
\alias{.check_data}
\title{.check_data}
\usage{
.check_data(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe (in case of no errors)
}
\description{
Performs a series of checks on a dataframe. For internal use. It returns an
error if some of the checks are not passed, otherwise return the \code{data}
argument.
}
\keyword{internal}
