% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_sigma}
\alias{.check_sigma}
\title{check_sigma}
\usage{
.check_sigma(nsim = 1000, n = 100, Sigma, skew = 0, kurt = 0)
}
\arguments{
\item{nsim}{number of simulations}

\item{n}{sample size}

\item{Sigma}{the variance-covariance matrix}

\item{skew}{skewness}

\item{kurt}{kurtosis}
}
\description{
Check how many times Sigma produce error when simulating data
}
\keyword{internal}
