% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.flatten_udata}
\alias{.flatten_udata}
\title{.flatten_udata}
\usage{
.flatten_udata(x)
}
\arguments{
\item{x}{a list}
}
\value{
the flattened version of the list without the \code{udata.} prefix
}
\description{
Flatten the \code{params} list when using \code{run_simulation()} with \code{udata}
}
\keyword{internal}
