% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_inds}
\alias{.get_inds}
\title{.get_inds}
\usage{
.get_inds(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Get all indicators from a dataset as columns starting with "x" and ending with
a number. Only for internal use.
}
\keyword{internal}
