% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_sim_pars}
\alias{.get_sim_pars}
\title{.get_sim_pars}
\usage{
.get_sim_pars(sargs)
}
\arguments{
\item{sargs}{list of simulation parameters.}
}
\value{
a list
}
\description{
Capture simulation parameters as function arguments from \code{...} or \code{udata} when
using the \code{run_simulation} function. For internal use.
}
\keyword{internal}
