% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_vcov}
\alias{.get_vcov}
\title{.get_vcov}
\usage{
.get_vcov(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a list
}
\description{
Compute the variance-covariance and correlation matrix. All columns are assumed to be numerical
and will be included in the matrix.
}
\keyword{internal}
