% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.udata_as_params}
\alias{.udata_as_params}
\title{.udata_as_params}
\usage{
.udata_as_params(x)
}
\arguments{
\item{x}{a list}
}
\value{
a list
}
\description{
Uniform the simulation result structure with and without \code{udata}
}
\keyword{internal}
