% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.within_time}
\alias{.within_time}
\title{.within_time}
\usage{
.within_time(start, maxt)
}
\arguments{
\item{start}{the starting time. A \code{POSIXct} \code{POSIXt} object returned by
calling \code{Sys.time()}.}

\item{maxt}{the maximum duration in seconds.}
}
\value{
logical
}
\description{
Internal function to check the simulation status. Every time the function is called
compare the current time with the simulation starting time. If the elapsed time is
beyond the maximum time the function return \code{FALSE}, otherwise \code{TRUE}.
For internal use.
}
\keyword{internal}
