% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_sigma}
\alias{gen_sigma}
\title{Generate a variance-covariance matrix}
\usage{
gen_sigma(p, rmin, rmax = NULL)
}
\arguments{
\item{p}{number of variables/indicators}

\item{rmin}{minimum correlation}

\item{rmax}{maximum correlation. Default to \code{NULL}. If \code{NULL} the
minimum is used thus using a single value.}
}
\value{
a matrix
}
\description{
Generate a random variance-covariance matrix sampling the correlation values
from a uniform distribution. The function does not guarantee that the matrix
is positive definite. The variance is fixed to 1 for simulating standardized data.
}
\examples{
gen_sigma(3, 0.5) # compound symmetry
gen_sigma(5, 0.5, 0.7)
}
