% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_d_from_D}
\alias{get_d_from_D}
\title{get_d_from_D}
\usage{
get_d_from_D(D, Sigma)
}
\arguments{
\item{D}{the desired multivariate \code{D} value}

\item{Sigma}{the variance-covariance matrix}
}
\value{
the \code{d} value to be used in all dimensions.
}
\description{
Find the \code{d} that when used in all dimensions give the desired \code{D}
value. It use the \code{optimize} function using several \code{d} values and
minimizing the difference with the desired \code{D}.
}
\examples{
D <- 2
S <- gen_sigma(3, 0.4)
d <- get_d_from_D(D, S)
d
multiD(d, S)
}
