% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{get_summary_stats}
\usage{
get_summary_stats(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
Compute all relevant summary statistics given a dataframes of all numerical columns.
}
\examples{
X <- sim_clust(2, 100, dmin = 0.5, rmin = 0.3, nind = 10)
get_summary_stats(X)
}
