% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kmeans_clust}
\alias{kmeans_clust}
\title{kmeans_clust}
\usage{
kmeans_clust(data, cmin = 2, cmax = 5, alpha = 0.05, debug = FALSE)
}
\arguments{
\item{data}{a dataframe with only the indicators to be used in the clustering
as columns.}

\item{cmin}{the minimum number of cluster to test. Default to 2}

\item{cmax}{the maximum number of clusters to test. Default to 5. Be careful that
increasing \code{cmax} will greatly increase the computation time.}

\item{alpha}{the alpha value for the Duda-Hart test. Default to \code{0.05}}

\item{debug}{logical indicating if silhouette values for each model should be returned}
}
\value{
the number of estimated clusters
}
\description{
Estimate the number of clusters using the \code{K-Means clustering} selecting the
best solutions using the silhouette value. Before finding the best solutions, the
Duda-Hart test is used to check if there are more at least two clusters.
}
\examples{
X <- sim_clust(2, 100, dmin = 0.5, rmin = 0.3, nind = 10)
X <- X[, 1:(ncol(X) - 1)] # excluding the last column
kmeans_clust(X)
}
