% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_based_clust}
\alias{model_based_clust}
\title{model_based_clust}
\usage{
model_based_clust(data, criteria = "BIC", cmin = 1, cmax = 5)
}
\arguments{
\item{data}{a dataframe with only the indicators to be used in the clustering
as columns.}

\item{criteria}{which criteria to select the best number of clusters. One of
\code{BIC} or \code{lrt} (likelihood ratio test).}

\item{cmin}{the minimum number of cluster to test. Default to 1}

\item{cmax}{the maximum number of clusters to test. Default to 5. Be careful that
increasing \code{cmax} will greatly increase the computation time.}
}
\value{
the estimated number of clusters
}
\description{
Estimate the number of clusters using the \code{mclust::Mclust()} function.
}
\examples{
X <- sim_clust(2, 100, dmin = 0.5, rmin = 0.3, nind = 10)
X <- X[, 1:(ncol(X) - 1)] # excluding the last column
model_based_clust(X)
}
