% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiD}
\alias{multiD}
\title{multiD}
\usage{
multiD(d, Sigma)
}
\arguments{
\item{d}{the vector of effect sizes. If \code{length(d)} is 1, the value is used for each dimension (indicator).}

\item{Sigma}{the variance-covariance matrix}
}
\value{
the Mahalanobis distance
}
\description{
Calculate the multivariate \code{D} based on the Mahalanobis distance as described in Del-Giudice2009-wn. This can be interpret as the
distance between the clusters on the n-dimensional space where \code{n} is the number of indicators.
}
\examples{
S <- gen_sigma(4, 0.7)
multiD(c(0.4, -0.3, 0.5, 0), S)
}
