% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_user_data}
\alias{prep_user_data}
\title{Summarize a dataframe extracting relevant statistics for the simulation}
\usage{
prep_user_data(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a list
}
\description{
The function takes a dataframe and return a list of properties used in the simulation, the \code{udata} object. In particular, it returns the number of observation, number of indicators,
summary statistics, the variance-covariance and correlation matrices.
}
\examples{
X <- sapply(1:10, function(x) rnorm(100, 0, 1))
X <- data.frame(X)
prep_user_data(X)

}
