% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustsim-methods.R
\name{print.clustsim}
\alias{print.clustsim}
\title{print.clustsim}
\usage{
\method{print}{clustsim}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print.clustsim
}
