% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quartiles}
\alias{quartiles}
\title{quartiles}
\usage{
quartiles(x, probs = c(0.25, 0.5, 0.75))
}
\arguments{
\item{x}{a numeric vector}

\item{probs}{the quantiles to be computed. Default to \code{c(0.25, 0.5, 0.75)}.}
}
\value{
a named numeric vector. Quantiles are expressed in percentage.
}
\description{
compute the 1st, 2nd (median) and 3rd quartiles.
}
\examples{
x <- rnorm(100)
quartiles(x)
}
