% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(
  niter,
  maxt = NULL,
  type = c("power", "type1"),
  method = c("mclust", "kmeans"),
  cmax = 5,
  dmin = 0,
  dmax = NULL,
  udata = NULL,
  progress = TRUE,
  updateProgress = NULL,
  ...
)
}
\arguments{
\item{niter}{number of iterations.}

\item{maxt}{the maximum time. If the number of iterations is not finished within the maximum time the simulations ends.}

\item{type}{character indicating the simulation type. One of \code{type1} or \code{power}.}

\item{method}{character indicating the method to use for estimating the number of cluster.
One of \code{mclust} or \code{kmeans}. \code{mclust} will call the \code{model_based_clust()} function and \code{kmeans}
the \code{kmeans_clust()} function.}

\item{cmax}{the maximum number of clusters to test. Default to 5. Be careful that
increasing \code{cmax} will significantly increase the computation time.}

\item{dmin}{the lower bound for the range of effect sizes (in ~Cohen's d like scale).}

\item{dmax}{the lower bound for the range of effect sizes (in ~Cohen's d like scale). If not specified, \code{dmax} takes the
value of \code{dmin} thus the effect size is the same for all indicators.}

\item{udata}{a list. The output of the \code{prep_user_data()} function. Default to \code{NULL}. If \code{udata} is not
\code{NULL} all the parameters passed to \code{...} are ignored.}

\item{progress}{logical indicating whether a progress bar (using the \code{progress} package) should be displayed.}

\item{updateProgress}{a function used fo displaying a progress bar in the Shiny app. Default to \code{NULL}}

\item{...}{named arguments that will be passed to the \code{sim_clust()} function. If \code{udata} is \code{null}, these
parameters are used for the simulation.}
}
\value{
a list containing the estimated power or type-1 error, the number of effective iterations, the proportion of
iterations with errors, a vector with the estimated number of clusters and a list with all simulation parameters.
}
\description{
Function to run the simulation for estimating the type-1 error or the statistical power
given a set of parameters.
}
\examples{
\dontrun{
run_simulation(niter = 100, maxt = 50, type = "type1", method = "mclust", 
nclust = 2, n = 100, dmin = 0.5, rmin = 0.5, nind = 3)
}
}
