% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_clust}
\alias{sim_clust}
\title{Simulate (un)clustered data assuming multivariate non-normal distributions}
\usage{
sim_clust(
  nclust = 1,
  n = NULL,
  dmin = 0,
  dmax = NULL,
  rmin = NULL,
  rmax = NULL,
  nind = NULL,
  skewmin = 0,
  skewmax = NULL,
  kurtmin = 0,
  kurtmax = NULL,
  pn1 = 0.5,
  debug = FALSE,
  udata = NULL
)
}
\arguments{
\item{nclust}{The number of clusters. Can be 1 or 2, default to 1}

\item{n}{The total number of observations. If \code{nclust} is 1, \code{n} is the number of observations,
if \code{nclust} is 2, \code{n * (1 - pn1)} is the number of observations for the cluster 2.}

\item{dmin}{the lower bound for the range of effect sizes (in ~Cohen's d like scale).}

\item{dmax}{the lower bound for the range of effect sizes (in ~Cohen's d like scale). If not specified, \code{dmax} takes the
value of \code{dmin} thus the effect size is the same for all indicators.}

\item{rmin}{the lower bound for the range of correlations between indicators. TODO The correlation is assumed to be the same within each cluster.}

\item{rmax}{the upper bound for the range of correlations between indicators. If not specified, \code{rmax} takes the value of
\code{rmin} thus using a single value.}

\item{nind}{the number of indicators}

\item{skewmin}{the lower bound for the range of skewness.}

\item{skewmax}{the upper bound for the range of skewness. If not specified, \code{skewmax} takes the value of \code{skewmin} thus using a single value.}

\item{kurtmin}{the lower bound for the range of kurtosi}

\item{kurtmax}{the upper bound for the range of kurtosi. If not specified, \code{kurtmax} takes the value of \code{kurtmin} thus using a single value.}

\item{pn1}{the proportion of observations belonging to the 1 cluster. Default to \code{0.5}.}

\item{debug}{logical indicating if returning the vectors of true values when sampling from a range.}

\item{udata}{list of summary statistics to use in the simulation. When specified, previous arguments are ignored.}
}
\value{
a dataframe with the simulated values
}
\description{
Simulate 1 or 2 clusters assuming multivariate non-normal distributions (specifying correlations, skewness, kurtosis, and the number of indicators) and the degree of separation (effect size) in standard deviation units (standardized data).
The function can works in two different ways.
\itemize{
\item When \code{udata} is \code{NULL}: need to be specified or left to the default value
\item When \code{udata} is provided: parameters are taken from \code{udata} that is the result of calling the \link{prep_user_data} function.
}
}
\examples{
# simulating with parameters
X <- sim_clust(nclust = 2, n = 100, dmin = 2, rmin = 0.5, nind = 5)
X
# with external data
udata <- prep_user_data(X)
sim_clust(udata = udata)
}
