% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_data}
\alias{sim_data}
\title{Simulate a dataset from a multivariate non-normal distribution}
\usage{
sim_data(n, d, Sigma, skew = NULL, kurt = NULL)
}
\arguments{
\item{n}{number of observations}

\item{d}{the effect size (i.e., the vector of means of the multivariate distribution). If \code{d} is a vector
of length 1, the value is recycled creating a vector of length \code{nrow(Sigma)}, otherwise \code{d} need to
be a vector of length \code{Sigma}. The function randomize the sign of \code{d} for each dimension.}

\item{Sigma}{the variance-covariance matrix}

\item{skew}{vector of skewnesses}

\item{kurt}{vector of kurtoses}
}
\value{
a dataframe with simulated data
}
\description{
Simulate a dataset coming from a multivariate non-normal distribution specifying the vector of means, variance-covariance matrix, vector of skewnesses and kurtoses. The function use the \code{\link[semTools:mvrnonnorm]{semTools::mvrnonnorm()}} function.
}
\examples{
Sigma <- gen_sigma(4, 0.5)
sim_data(100, 0.5, Sigma)
}
